Use more legit approach to install a phaser
It still relies upon compiler internals, the World object in particular
and the way a unit is compiled. But otherwise the code is currently
using `Block` `add_phaser` method and relies on the fact that compunit
body is a block which `ENTER` phaser is executed when unit's bytecode is
loaded and invoked. This works reliably for both module and
script/application code.