Improve parsing of paths and value-only blocks
Allow `dir1/dir2/file` form of paths to be used as a block value without
the grammar trying to consider `dir1` as an option.

Overall be more specific about what could be an option name by
restricting it to be only followed by a whitespace of a kind or a
statement terminator.