Fix a bug where non-values could be parsed as a value
In particular, races for parsing comments or options could be win in
favor of value.

Also, bare file paths looking like options, i.e. starting with a keyword
alike followed by a white space or statement terminator, are now
prohibited unless used as option value or as part of a value-only block.